+++ src/main/java/vn/hactanco/socialnetwork/model/User.java
@@ -1,5 +1,6 @@
 package vn.hactanco.socialnetwork.model;
 
+import java.io.Serializable;
 import java.time.Instant;
 import java.util.List;
 
@@ -25,8 +26,12 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class User {
+public class User implements Serializable {
 
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 1L;
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;